requirements.txt
# python-dotenv~=1.0.1
filetype~=1.2.0
python-telegram-bot~=21.6
